backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
anthropic==0.51
pydantic==2.6.0
python-multipart==0.0.20
python-dotenv==1.0.1
manim==0.19.0
boto3==1.34.69
groq==0.4.1